opet_app.py
from flask import Flask, request, jsonify
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification
import concurrent.futures
import fasttext
# Assume `spacy_udpipe` and other necessary imports are done earlier

app = Flask(__name__)

# Preload models (assuming functions to load tokenizers/models are defined)
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# Load sentiment models onto GPU
ur_sent_model, ru_sent_model, en_sent_model = load_sentiment_models(device=device)
lang_model = fasttext.load_model("path/to/your/fasttext/model.bin")

@app.route('/predict', methods=['POST'])
def predict():
    data = request.get_json(force=True)
    text = data.get('text', '')

    # Run tasks that don't require GPU in parallel
    with concurrent.futures.ThreadPoolExecutor() as executor:
        lang_future = executor.submit(detect_language, text, lang_model)
        actionable_future = executor.submit(find_actionables, text)
        profanity_future = executor.submit(check_profanity, text)

    detected_language = lang_future.result()
    sentiment = analyze_sentiment(text, detected_language, device)

    # Wait for other tasks to complete
    actionables = actionable_future.result()
    profanities = profanity_future.result()

    # Post-processing (e.g., adjusting sentiment based on profanities)
    if profanities:
        sentiment = "Negative"

    response = {
        "Detected Language": detected_language,
        "Sentiment": sentiment,
        "Actionable words": actionables,
        "Profanities": profanities,
    }

    return jsonify(response)

def load_sentiment_models(device):
    # Placeholder: load your sentiment analysis models here and move them to the device
    return ur_sent_model, ru_sent_model, en_sent_model

def detect_language(text, lang_model):
    # Placeholder: implement your language detection logic here
    return "eng_Latn"

def find_actionables(text):
    # Placeholder: implement logic to find actionable words
    return []

def check_profanity(text):
    # Placeholder: implement logic to check for profanity
    return []

def analyze_sentiment(text, detected_language, device):
    # Placeholder: Use the correct model based on detected_language
    # Prepare the input, perform the model inference, and return the sentiment
    return "Positive"

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000)
