requirements.txt
transformers
spacy_udpipe
torch
pandas
flask
gunicorn
fasttext-wheel

